backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
openai
python-pptx
reportlab
jinja2
python-dotenv
resend
stripe
groq
playwright